+++ config/update.js
@@ -3,13 +3,13 @@ const https = require('https');
 const { src, dest, task, series } = require('gulp');
 const $replace = require('gulp-replace');
 
-// updated: 2020.07.04
+// updated: 2020.08.26
 const LATEST_VERSIONS = {
-  filled: 53,
-  outlined: 22,
-  round: 22,
-  twoTone: 21,
-  sharp: 23
+  filled: 55,
+  outlined: 25,
+  round: 24,
+  twoTone: 23,
+  sharp: 25
 };
 
 // Update Material Components Web for BalmUI
@@ -185,13 +185,14 @@ function updateMDITask(cb) {
     .on('error', (e) => {
       console.error(e);
     });
+
   cb();
 }
 
 task('update:mdi', updateMDITask);
 
 // Set Material Icons Category
-// const MDI_JSON = 'https://fonts.google.com/metadata/icons';
+// const MDI_JSON = 'https://fonts.google.com/metadata/icons'; // NOTE: manual download `json.txt`
 const sourceData = './docs/data/txt.json';
 const targetData = './docs/data/icons.json';